# backend/scripts/swagger_v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "DataLabellingSupportSystem.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/annotator/tasks/{taskId}/start": {
      "post": {
        "tags": [
          "Annotator"
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Boolean"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Boolean"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/annotator/tasks/{taskId}/items": {
      "get": {
        "tags": [
          "Annotator"
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Collections.Generic.List_DataLabellingSupportSystem.Api.DTOs.Responses.Annotator.AnnotatorTaskItemResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Collections.Generic.List_DataLabellingSupportSystem.Api.DTOs.Responses.Annotator.AnnotatorTaskItemResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Collections.Generic.List_DataLabellingSupportSystem.Api.DTOs.Responses.Annotator.AnnotatorTaskItemResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/annotator/tasks/{taskId}/labels": {
      "get": {
        "tags": [
          "Annotator"
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Collections.Generic.List_DataLabellingSupportSystem.Api.DTOs.Responses.Annotator.LabelResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Collections.Generic.List_DataLabellingSupportSystem.Api.DTOs.Responses.Annotator.LabelResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Collections.Generic.List_DataLabellingSupportSystem.Api.DTOs.Responses.Annotator.LabelResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/annotator/tasks/{taskId}/guideline": {
      "get": {
        "tags": [
          "Annotator"
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Annotator.ProjectGuidelineResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Annotator.ProjectGuidelineResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Annotator.ProjectGuidelineResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/annotator/tasks/{taskId}/data-item/content": {
      "get": {
        "tags": [
          "Annotator"
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/annotator/tasks/{taskId}/annotations": {
      "get": {
        "tags": [
          "Annotator"
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Collections.Generic.List_DataLabellingSupportSystem.Api.DTOs.Responses.Annotator.AnnotatorAnnotationResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Collections.Generic.List_DataLabellingSupportSystem.Api.DTOs.Responses.Annotator.AnnotatorAnnotationResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Collections.Generic.List_DataLabellingSupportSystem.Api.DTOs.Responses.Annotator.AnnotatorAnnotationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/annotator/tasks/{taskId}/annotations/draft": {
      "put": {
        "tags": [
          "Annotator"
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Annotator.UpsertTaskItemAnnotationsRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Annotator.UpsertTaskItemAnnotationsRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Annotator.UpsertTaskItemAnnotationsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Boolean"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Boolean"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/annotator/tasks/{taskId}/annotations/submit": {
      "post": {
        "tags": [
          "Annotator"
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Annotator.UpsertTaskItemAnnotationsRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Annotator.UpsertTaskItemAnnotationsRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Annotator.UpsertTaskItemAnnotationsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Boolean"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Boolean"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/annotator/tasks/{taskId}/ai-suggest": {
      "post": {
        "tags": [
          "Annotator"
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "apply",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Annotator.AiAssistSuggestResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Annotator.AiAssistSuggestResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Annotator.AiAssistSuggestResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tasks": {
      "get": {
        "tags": [
          "LabelingTasks"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "LabelingTasks"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Models.LabelingTask"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Models.LabelingTask"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Models.LabelingTask"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "LabelingTasks"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Models.LabelingTask"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Models.LabelingTask"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Models.LabelingTask"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "LabelingTasks"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Annotator.AiAssistSuggestResponse": {
        "type": "object",
        "properties": {
          "isSuccess": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Responses.Annotator.AiAssistSuggestResponse"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Annotator.ProjectGuidelineResponse": {
        "type": "object",
        "properties": {
          "isSuccess": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Responses.Annotator.ProjectGuidelineResponse"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Boolean": {
        "type": "object",
        "properties": {
          "isSuccess": {
            "type": "boolean"
          },
          "data": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Collections.Generic.List_DataLabellingSupportSystem.Api.DTOs.Responses.Annotator.AnnotatorAnnotationResponse": {
        "type": "object",
        "properties": {
          "isSuccess": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Responses.Annotator.AnnotatorAnnotationResponse"
            },
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Collections.Generic.List_DataLabellingSupportSystem.Api.DTOs.Responses.Annotator.AnnotatorTaskItemResponse": {
        "type": "object",
        "properties": {
          "isSuccess": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Responses.Annotator.AnnotatorTaskItemResponse"
            },
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Collections.Generic.List_DataLabellingSupportSystem.Api.DTOs.Responses.Annotator.LabelResponse": {
        "type": "object",
        "properties": {
          "isSuccess": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Responses.Annotator.LabelResponse"
            },
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DataLabellingSupportSystem.Api.DTOs.Requests.Annotator.UpsertTaskItemAnnotationsRequest": {
        "type": "object",
        "properties": {
          "objects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Annotator.UpsertAnnotationObject"
            },
            "nullable": true
          },
          "predictionId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DataLabellingSupportSystem.Api.DTOs.Responses.Annotator.AnnotatorTaskItemResponse": {
        "type": "object",
        "properties": {
          "taskId": {
            "type": "string",
            "nullable": true
          },
          "dataItemId": {
            "type": "string",
            "nullable": true
          },
          "storageProvider": {
            "type": "string",
            "nullable": true
          },
          "objectKey": {
            "type": "string",
            "nullable": true
          },
          "originalWidth": {
            "type": "integer",
            "format": "int32"
          },
          "originalHeight": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "DataLabellingSupportSystem.Api.DTOs.Responses.Annotator.LabelResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "yoloClassId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "DataLabellingSupportSystem.Api.DTOs.Responses.Annotator.ProjectGuidelineResponse": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "string",
            "nullable": true
          },
          "guideline": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Paste only the JWT access token here. Swagger will send: Authorization: Bearer <token>",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}